Refactor import statements to use the new product package structure for DTOs
