+++ src/main/java/es/marcha/backend/mapper/ProductMapper.java
@@ -3,7 +3,7 @@ package es.marcha.backend.mapper;
 import java.util.List;
 
 import es.marcha.backend.dto.request.ecommerce.ProductRequestDTO;
-import es.marcha.backend.dto.response.ecommerce.ProductResponseDTO;
+import es.marcha.backend.dto.response.ecommerce.product.ProductResponseDTO;
 import es.marcha.backend.model.ecommerce.Subcategory;
 import es.marcha.backend.model.ecommerce.product.Product;
